app/requirements.txt
celery==4.4.1
Click==7.0
Cython==0.29.15
Flask==1.1.1
Flask-And-Redis==1.0.0
Flask-Bcrypt==0.7.1
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.13.2
kombu==4.6.8
MarkupSafe==1.1.1
pylint==2.4.4
PyMySQL==0.9.3
python-dotenv==0.12.0
pytz==2019.1
redis==3.4.1
requests==2.22.0
sqlparse==0.3.0
Werkzeug==1.0.0